requirements.txt
torch~=2.3.1
pandas~=2.2.2
pillow~=10.3.0
torchvision~=0.18.1
matplotlib~=3.9.0
tqdm~=4.66.4